[FIX] changes in reduced calendar
