+++ pms_pwa/__manifest__.py
@@ -26,7 +26,7 @@
         "templates/dashboard.xml",
         "templates/calendar.xml",
         "templates/calendar_config.xml",
-        "templates/reduced_calendar.xml",
+        # "templates/reduced_calendar.xml",
         "templates/login.xml",
         "templates/rooms.xml",
         "templates/rates.xml",
