+++ pms_pwa/templates/_includes/head.xml
@@ -36,11 +36,11 @@
                 type="text/scss"
                 href="/pms_pwa/static/src/scss/o_pms_pwa_calendar.scss"
             />
-            <link
+            <!-- <link
                 rel="stylesheet"
                 type="text/scss"
                 href="/pms_pwa/static/src/scss/o_pms_pwa_reduced_calendar.scss"
-            />
+            /> -->
             <link
                 rel="stylesheet"
                 type="text/scss"
@@ -65,7 +65,7 @@
             <link rel="stylesheet" href="/pms_pwa/static/src/scss/vendor/bs-stepper.min.css"/>
             <link rel="stylesheet" href="/pms_pwa/static/src/scss/vendor/daterangepicker.css"/>
             <link rel="stylesheet" href="/pms_pwa/static/src/scss/vendor/bootstrap-select.min.css" />
-            <link rel="stylesheet" href="/pms_pwa/static/src/scss/vendor/print.min.css" />
+            <!-- <link rel="stylesheet" href="/pms_pwa/static/src/scss/vendor/print.min.css" /> -->
         </xpath>
     </template>
     <template id="website_roomdoo_js" inherit_id="website.assets_frontend">
@@ -100,10 +100,10 @@
                 type="text/javascript"
                 src="/pms_pwa/static/src/js/o_pms_pwa_calendar_config.js"
             />
-            <script
+            <!-- <script
                 type="text/javascript"
                 src="/pms_pwa/static/src/js/o_pms_pwa_reduced_calendar.js"
-            />
+            /> -->
             <script
                 type="text/javascript"
                 src="/pms_pwa/static/src/js/o_pms_pwa_notify.js"
@@ -119,10 +119,10 @@
             <script src="/web/static/lib/Chart/Chart.js"></script>
 
             <script src="/pms_pwa/static/src/js/vendor/bootstrap-select.min.js"></script>
-            <script
+            <!-- <script
                 type="text/javascript"
                 src="/pms_pwa/static/src/js/vendor/print.min.js"
-            />
+            /> -->
         </xpath>
     </template>
 </odoo>
